# feat: enable gpu async channel feature to improve startup