@deepak1556 2025-02-23
I will push the flags into VSCode by default from March milestone since it is too late in the current milestone. However, I would like to start testing with the flag in the perf bot.

Here is the upstream CL https://chromium-review.googlesource.com/c/chromium/src/+/6094045 that enables it since 134, we should get it once we update to E35