@deepak1556 2025-02-23
Closing in favor of https://github.com/microsoft/vscode/pull/241640